+++ lib/models/src/business_unit.dart
@@ -6,6 +6,7 @@ part 'business_unit.g.dart';
 @freezed
 class BusinessUnit with _$BusinessUnit {
   const factory BusinessUnit({
+    int? id,
     String? name,
     DateTime? createdDate,
   }) = _BusinessUnit;