+++ lib/models/src/business_unit.g.dart
@@ -8,6 +8,7 @@ part of 'business_unit.dart';
 
 _$BusinessUnitImpl _$$BusinessUnitImplFromJson(Map<String, dynamic> json) =>
     _$BusinessUnitImpl(
+      id: (json['id'] as num?)?.toInt(),
       name: json['name'] as String?,
       createdDate: json['createdDate'] == null
           ? null
@@ -16,6 +17,7 @@ _$BusinessUnitImpl _$$BusinessUnitImplFromJson(Map<String, dynamic> json) =>
 
 Map<String, dynamic> _$$BusinessUnitImplToJson(_$BusinessUnitImpl instance) =>
     <String, dynamic>{
+      'id': instance.id,
       'name': instance.name,
       'createdDate': instance.createdDate?.toIso8601String(),
     };